静的ページ追加、theme.jsのなかにtheme.stylesと別に定義すると、ページ内で<Styled.h1>などとして利用できる（https://qiita.com/SotaSuzuki/items/f2993afcaf2bdd438c29）
